# Use new value! and predict! methods of Emipircal risks to reduce memory allocation