@lindahua 2015-09-10
The new version of EmpiricalRisks with `predict!` has been tagged:

https://github.com/JuliaLang/METADATA.jl/pull/3329